#201: Rename the interface to Std, the implementation to StdOutput
